Some development features added:
When developing for fermentrack using a fork of fermentrack
this patch makes it possible to specify with repository to
use (with -r) and branch (-b). I was not sure if this should
be documented or not but I have added it to the help output
with a note when it is used.